@liggitt 2017-07-10
@k8s-bot ok to test
lgtm, but would like a sanity check from @smarterclayton