@php-coder 2017-07-10
The error
```
init_test.go:522: [0] unexpected error: dial tcp [::]:42467: connect: cannot assign requested address
init_test.go:522: [1] unexpected error: dial tcp [::]:37453: connect: cannot assign requested address
init_test.go:522: [2] unexpected error: dial tcp [::]:41835: connect: cannot assign requested address
init_test.go:522: [3] unexpected error: dial tcp [::]:39743: connect: cannot assign requested address
init_test.go:522: [4] unexpected error: dial tcp [::]:41001: connect: cannot assign requested address
init_test.go:522: [5] unexpected error: dial tcp [::]:34457: connect: cannot assign requested address
init_test.go:511: [6] unexpected error: want x509.HostnameError, got: *net.OpError
init_test.go:511: [7] unexpected error: want x509.HostnameError, got: *net.OpError
init_test.go:511: [8] unexpected error: want x509.HostnameError, got: *net.OpError
init_test.go:511: [9] unexpected error: want x509.HostnameError, got: *net.OpError
init_test.go:515: [10] unexpected error: want x509.UnknownAuthorityError, got: *net.OpError
init_test.go:515: [11] unexpected error: want x509.UnknownAuthorityError, got: *net.OpError
init_test.go:511: [12] unexpected error: want x509.HostnameError, got: *net.OpError
init_test.go:511: [13] unexpected error: want x509.HostnameError, got: *net.OpError
```
shouldn't happen again because it was just fixed: https://github.com/kubernetes/kubernetes/issues/47304