@php-coder 2017-07-10
Another one is something new:
```
I0710 02:05:25.773] +++ [0710 02:05:25] Verifying Prerequisites....
W0710 02:05:25.874] + make release
I0710 02:05:26.026] +++ [0710 02:05:26] Building Docker image kube-build:build-9a111c413b-5-v1.8.3-1
Build timed out (after 90 minutes). Marking the build as failed.
Build was aborted
```
I don't know what is it and I wasn't able to find out a reason.