@php-coder 2017-07-10
/test pull-kubernetes-unit
/test pull-kubernetes-e2e-kops-aws