@smarterclayton 2017-07-17
I think this looks correct, although did this change the benchmark performance?