@liggitt 2017-08-18
@php-coder did you check the benchmark times? this looks ready to merge otherwise