@php-coder 2017-08-25
Here is the benchmark results:

| Test | Number of cycles | Before (avg after 5 runs, ns/op) | After (avg after 5 runs, ns/op) |
| --- | --- | --- | --- |
| BenchmarkGCMRead_16_1024-4          	| 1000000	   | 1093.4 | 1076.6 |
| BenchmarkGCMRead_32_1024-4          	| 1000000	   | 1256.4 | 1246.4 |
| BenchmarkGCMRead_32_16384-4         	 | 100000	  | 11905.8 | 11904.0 |
| BenchmarkGCMRead_32_16384_Stale-4   	 | 100000	  | 12071.6 | 11977.6 |
| BenchmarkCBCRead_32_1024-4          	 | 500000	   | 3629.4 | 3648.4 |
| BenchmarkCBCRead_32_16384-4         	 |  30000	  | 52878.0 | 53340.2 |
| BenchmarkCBCRead_32_16384_Stale-4   	 |  30000	  | 53399.0 | 52826.2 |
| BenchmarkSecretboxRead_32_1024-4          	 | 500000	   | 2536.2 | 2566.2 |
| BenchmarkSecretboxRead_32_16384-4         	  | 50000	  | 29687.4 | 29647.2 |
| BenchmarkSecretboxRead_32_16384_Stale-4   	  | 50000	  | 29813.2 | 29802.0 |

As far I can see the results mostly the same.