@php-coder 2017-08-25
/test pull-kubernetes-e2e-kops-aws